refactor (front): Move some Vue components to new 'views' directory
This commit refactors the project structure by creating a new 'views' directory at the root of the project, and moving some Vue component files from the 'components' directory to the new location. This change improves the organization of the project, separating components used for building views from other types of components.